refactor: Export cells component from File into cells folder
in seperate files for more clarity
