+++ src/drive/web/modules/filelist/File.spec.jsx
@@ -10,7 +10,7 @@ import { render, fireEvent } from '@testing-library/react'
 import { createMockClient } from 'cozy-client'
 
 import AppLike from 'test/components/AppLike'
-import { DumbFile, getParentLink } from './File'
+import { DumbFile } from './File'
 import { folder, actionsMenu } from 'test/data'
 
 const client = createMockClient({})
@@ -45,29 +45,6 @@ const setup = ({
   )
   return { root }
 }
-
-describe('getParentLink function', () => {
-  it('should return the first link in the element ancestors', () => {
-    const div = document.createElement('div')
-    const link = document.createElement('a')
-    link.className = 'my-link'
-    const span = document.createElement('span')
-    const span2 = document.createElement('span')
-    div.appendChild(link)
-    link.appendChild(span)
-    span.appendChild(span2)
-    const result = getParentLink(span2)
-    expect(result).toEqual(link)
-  })
-
-  it('should return null if there is no link in the element ancestors', () => {
-    const div = document.createElement('div')
-    const span = document.createElement('span')
-    div.appendChild(span)
-    const result = getParentLink(span)
-    expect(result).toBeNull()
-  })
-})
 
 describe('File', () => {
   describe('default behavior', () => {
