+++ src/drive/web/modules/filelist/FileOpener.spec.js
@@ -1,0 +1,24 @@
+import { getParentLink } from './FileOpener'
+
+describe('getParentLink function', () => {
+  it('should return the first link in the element ancestors', () => {
+    const div = document.createElement('div')
+    const link = document.createElement('a')
+    link.className = 'my-link'
+    const span = document.createElement('span')
+    const span2 = document.createElement('span')
+    div.appendChild(link)
+    link.appendChild(span)
+    span.appendChild(span2)
+    const result = getParentLink(span2)
+    expect(result).toEqual(link)
+  })
+
+  it('should return null if there is no link in the element ancestors', () => {
+    const div = document.createElement('div')
+    const span = document.createElement('span')
+    div.appendChild(span)
+    const result = getParentLink(span)
+    expect(result).toBeNull()
+  })
+})
